Update Greeting so it does not flash
