+++ src/pages/index.js
@@ -3,12 +3,13 @@ import * as React from 'react'
 import Layout from '../components/layout'
 import SEO from '../components/seo'
 import { Greeting } from '../components'
+import { greetingMessage } from '../utils'
 
 const IndexPage = () => (
   <Layout>
     <SEO title='Home' />
     {/* <h1 className='text-xl2'>Greetings folks,</h1> */}
-    <Greeting />
+    <Greeting greetingMessage={greetingMessage()} />
     <p>Welcome to KoaMar 2020</p>
     <p>It's been 24 years, about time something happened!</p>
     <hr />
