+++ src/utils/greeting-message.js
@@ -1,0 +1,24 @@
+export const greetingMessage = () => {
+  const today = new Date()
+
+  // This should still give me the correct time in another country right?
+  // I just want to be able to read what it returns!
+  const currentHour = today.toLocaleTimeString('pt-PT').slice(0, 2)
+
+  // Starting with a simple implementation without regard to sunrise / sunset
+  // It would be nice if it took that into account
+  const timeOfDay = () => {
+    switch (true) {
+      case currentHour < '12':
+        return 'morning'
+      case currentHour < '18':
+        return 'afternoon'
+      case currentHour <= '24':
+        return 'evening'
+      default:
+        return 'night'
+    }
+  }
+
+  return `Good ${timeOfDay()} everyone,`
+}
